Split create into core and main
