+++ Cargo.toml
@@ -9,16 +9,15 @@ keywords = [ "crypto", "ECDSA", "secp256k1", "bitcoin", "no_std" ]
 edition = "2018"
 
 [dependencies]
+libsecp256k1-core = { version = "0.4.0", path = "core", default-features = false }
+arrayref = "0.3"
 rand = { version = "0.7", default-features = false }
+digest = "0.8"
+base64 = { version = "0.11.0", default-features = false }
 hmac-drbg = { version = "0.2", optional = true }
 sha2 = { version = "0.8", optional = true, default-features = false }
-digest = "0.8"
 typenum = { version = "1.11", optional = true }
-arrayref = "0.3"
-subtle = { version = "2.2", default-features = false }
-serde =  { version = "1.0.104", features = ["derive"], default-features = false }
-base64 = { version = "0.11.0", default-features = false }
-crunchy = "0.2"
+serde = { version = "1.0.104", features = ["derive"], default-features = false }
 
 [dev-dependencies]
 secp256k1-test = "0.7"
@@ -29,9 +28,8 @@ hex-literal = "0.2.1"
 
 [features]
 default = ["std", "hmac"]
-std = ["subtle/std", "rand/std", "sha2/std", "serde/std", "base64/std"]
+std = ["libsecp256k1-core/std", "sha2/std", "rand/std", "serde/std", "base64/std"]
 hmac = ["hmac-drbg", "sha2", "typenum"]
-noconst = []
 
 [workspace]
 members = [
