+++ core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "libsecp256k1-core"
+description = "Core functions for pure Rust secp256k1 implementation."
+license = "Apache-2.0"
+version = "0.4.0"
+authors = ["Wei Tang <hi@that.world>"]
+repository = "https://github.com/paritytech/libsecp256k1"
+keywords = [ "crypto", "ECDSA", "secp256k1", "bitcoin", "no_std" ]
+edition = "2018"
+
+[dependencies]
+subtle = { version = "2.2", default-features = false }
+crunchy = "0.2"
+digest = "0.8"
+
+[features]
+default = ["std"]
+std = ["subtle/std"]
