+++ core/src/ecmult.rs
