+++ core/src/scalar.rs
